Delete a print statement; fix that we were still leaking tasks
Previously we were overwriting the entire widget list for the panel for
every widget we initialize. This was a Bad Thing™ because it would leave
the widget watch tasks that are not the last one in the panel to
possibly get garbage collected and thus possibly cause the suspected
high CPU usage bugs we were seeing previously.