Fix audio weird and also some logging+multimonitor bugs
Note: audio is still weird since it's possible for pipewire to generate
a sink add event without changing the default on the server first, which
is absolutely broken and I am not sure how I am supposed to deal with
it. Besides perhaps asking the server every time there's a volume change
so it's at least easy to poke it back.